# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FactSnap API",
        "description": "API for managing educational categories, packages, and facts.\n *\n * # Authentication\n * \n * This API supports **dual authentication**:\n * \n * ## Firebase Authentication (for regular users)\n * - **Token Type**: Firebase ID Token\n * - **Format**: `Authorization: Bearer <firebase_id_token>`\n * - **Access**: 'free', 'loggedIn', and 'premium' packages (if premium user)\n * \n * ## Sanctum Authentication (for guest users)  \n * - **Token Type**: Sanctum Bearer Token (10-year expiration)\n * - **Format**: `Authorization: Bearer <sanctum_token>`\n * - **Access**: 'free' and 'loggedIn' packages ('premium' if created as premium guest)\n * \n * ## Package Access Levels\n * - **free**: No authentication required\n * - **loggedIn**: Requires authentication (Firebase OR Sanctum)\n * - **premium**: Requires premium user status",
        "contact": {
            "email": "support@factsnap.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Factsnap API Server"
        }
    ],
    "paths": {
...
        "/auth/firebase/sign-in": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign in with Firebase Phone Auth",
                "description": "Verifies a Firebase ID token (phone auth), upserts a local user, and returns the user profile. Returns 201 on first-time creation and 200 on subsequent sign-ins.",
                "operationId": "ee6d44947b2f720602062914e56e4b59",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idToken"
                                ],
                                "properties": {
                                    "idToken": {
                                        "description": "Firebase ID token obtained after phone authentication",
                                        "type": "string",
                                        "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created – First-time sign-in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User created"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK – Existing user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Signed in"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized – Invalid/revoked token or not a phone sign-in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiError"
                                        },
                                        {
                                            "properties": {
                                                "errors": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests – Rate limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Too Many Attempts."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get the current authenticated user",
                "description": "Returns the authenticated user profile. Requires a valid Firebase ID token in Authorization header.",
                "operationId": "2ade2b90c69c0f023057790461d6c328",
                "responses": {
                    "200": {
                        "description": "Current user profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized – Invalid/missing token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiError": {
                "properties": {
                    "code": {
                        "description": "Error code for programmatic handling",
                        "type": "integer",
                        "example": 1001
                    },
                    "message": {
                        "description": "User-friendly error message",
                        "type": "string",
                        "example": "The requested category could not be found."
                    },
                    "details": {
                        "description": "Technical details and context for developers",
                        "type": "string",
                        "example": "Category with UUID '123e4567-e89b-12d3-a456-426614174000' does not exist in the database."
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [
                    "id",
                    "name",
                    "is_guest",
                    "is_premium",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "firebase_uid": {
                        "description": "Firebase UID for Firebase authenticated users, null for guest users",
                        "type": "string",
                        "nullable": true
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+15551234567",
                        "nullable": true
                    },
                    "phone_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "is_guest": {
                        "description": "True for guest users created via Sanctum, false for Firebase users",
                        "type": "boolean"
                    },
                    "is_premium": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "firebaseAuth": {
                "type": "http",
                "description": "Firebase ID Token obtained from Firebase SDK. Use in Authorization header: Bearer <firebase_id_token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctumAuth": {
                "type": "http",
                "description": "Sanctum Bearer Token for guest users. Get token from POST /api/auth/guest. Use in Authorization header: Bearer <sanctum_token>",
                "bearerFormat": "Token",
                "scheme": "bearer"
            },
            "dualAuth": {
                "type": "http",
                "description": "Accepts either Firebase ID tokens OR Sanctum tokens. Firebase is tried first, then Sanctum as fallback.",
                "scheme": "bearer"
            }
        }
    },
...
}